+++ spec/www/spec/browser-check-startup.js
@@ -2,18 +2,11 @@
 
 var MYTIMEOUT = 12000;
 
-var isAndroid = /Android/.test(navigator.userAgent);
 var isWP8 = /IEMobile/.test(navigator.userAgent); // Matches WP(7/8/8.1)
-//var isWindows = /Windows NT/.test(navigator.userAgent); // Windows [NT] (8.1)
-var isWindows = /Windows /.test(navigator.userAgent); // Windows (8.1)
-//var isWindowsPC = /Windows NT/.test(navigator.userAgent); // Windows [NT] (8.1)
-//var isWindowsPhone_8_1 = /Windows Phone 8.1/.test(navigator.userAgent); // Windows Phone 8.1
-//var isIE = isWindows || isWP8 || isWindowsPhone_8_1;
-var isIE = isWindows || isWP8;
-var isWebKit = !isIE; // TBD [Android or iOS]
+var isWindows = /Windows /.test(navigator.userAgent); // Windows 8.1/Windows Phone 8.1/Windows 10
 
 window.hasBrowser = true;
-window.hasWebKitBrowser = isWebKit;
+window.hasBrowserWithWebSQL = !isWindows && !isWP8;
 
 describe('check startup', function() {
   it('receives deviceready event', function(done) {
@@ -24,7 +17,7 @@ describe('check startup', function() {
   }, MYTIMEOUT);
 
   it('has openDatabase', function() {
-    if (isWebKit) expect(window.openDatabase).toBeDefined();
+    if (window.hasBrowserWithWebSQL) expect(window.openDatabase).toBeDefined();
     expect(window.sqlitePlugin).toBeDefined();
     expect(window.sqlitePlugin.openDatabase).toBeDefined();
   });