+++ spec/www/spec/db-sql-operations-test.js
@@ -4,55 +4,24 @@ var MYTIMEOUT = 12000;
 
 var DEFAULT_SIZE = 5000000; // max to avoid popup in safari/ios
 
-// FUTURE TODO replace in test(s):
-function ok(test, desc) { expect(test).toBe(true); }
-function equal(a, b, desc) { expect(a).toEqual(b); } // '=='
-
-// XXX TODO REFACTOR OUT OF OLD TESTS:
-var wait = 0;
-var test_it_done = null;
-function xtest_it(desc, fun) { xit(desc, fun); }
-function test_it(desc, fun) {
-  wait = 0;
-  it(desc, function(done) {
-    test_it_done = done;
-    fun();
-  }, MYTIMEOUT);
-}
-function stop(n) {
-  if (!!n) wait += n
-  else ++wait;
-}
-function start(n) {
-  if (!!n) wait -= n;
-  else --wait;
-  if (wait == 0) test_it_done();
-}
-
+var isWP8 = /IEMobile/.test(navigator.userAgent); // Matches WP(7/8/8.1)
 var isWindows = /Windows /.test(navigator.userAgent); // Windows
 var isAndroid = !isWindows && /Android/.test(navigator.userAgent);
 
-// NOTE: In the common storage-master branch there is no difference between the
-// default implementation and implementation #2. But the test will also apply
-// the androidLockWorkaround: 1 option in the case of implementation #2.
-var scenarioList = [
+// NOTE: In certain versions such as Cordova-sqlcipher-adapter there is
+// no difference between the default implementation and implementation #2.
+// But the test will also specify the androidLockWorkaround: 1 option
+// in case of implementation #2 (also ignored by Cordova-sqlcipher-adapter).
+var pluginScenarioList = [
   isAndroid ? 'Plugin-implementation-default' : 'Plugin',
-  'HTML5',
   'Plugin-implementation-2'
 ];
 
-var scenarioCount = (!!window.hasWebKitBrowser) ? (isAndroid ? 3 : 2) : 1;
+var pluginScenarioCount = isAndroid ? 2 : 1;
 
 var mytests = function() {
 
-  describe('Plugin: plugin-specific sql test(s)', function() {
-
-    var pluginScenarioList = [
-      isAndroid ? 'Plugin-implementation-default' : 'Plugin',
-      'Plugin-implementation-2'
-    ];
-
-    var pluginScenarioCount = isAndroid ? 2 : 1;
+  describe('Plugin: plugin-specific SQL operations test(s)', function() {
 
     for (var i=0; i<pluginScenarioCount; ++i) {
 
@@ -93,39 +62,69 @@ var mytests = function() {
           return window.sqlitePlugin.openDatabase(dbopts, okcb, errorcb);
         }
 
-        test_it(suiteName + "Multiple db.executeSql string result test", function() {
+        it(suiteName + 'Inline db.executeSql US-ASCII String manipulation test with null parameter list', function(done) {
+          var db = openDatabase("Inline-db-sql-US-ASCII-string-test-with-null-parameter-list.db", "1.0", "Demo", DEFAULT_SIZE);
+          expect(db).toBeDefined();
+
+          db.executeSql("SELECT UPPER('Some US-ASCII text') AS uppertext", null, function(res) {
+            expect(res.rows.item(0).uppertext).toBe("SOME US-ASCII TEXT");
+
+            db.close(done, done);
+          });
+        }, MYTIMEOUT);
+
+        it(suiteName + 'Inline db.executeSql US-ASCII String manipulation test with empty ([]) parameter list', function(done) {
+          var db = openDatabase("Inline-db-sql-US-ASCII-string-test-with-empty-parameter-list.db", "1.0", "Demo", DEFAULT_SIZE);
+          expect(db).toBeDefined();
+
+          db.executeSql("SELECT UPPER('Some US-ASCII text') AS uppertext", [], function(res) {
+            expect(res.rows.item(0).uppertext).toBe("SOME US-ASCII TEXT");
+
+            db.close(done, done);
+          });
+        }, MYTIMEOUT);
+
+        it(suiteName + 'Multiple db.executeSql string result test', function() {
           // NOTE: this test checks that for db.executeSql(), the result callback is
           // called exactly once, with the proper result:
           var db = openDatabase("Multiple-DB-sql-String-result-test.db", "1.0", "Demo", DEFAULT_SIZE);
+          expect(db).toBeDefined();
 
-          var expected = [ 'FIRST', 'SECOND' ];
-          var i=0;
-
-          ok(!!db, 'valid db object');
-
-          stop(2);
-
-          var okcb = function(result) {
-            if (i > 1) {
-              ok(false, "unexpected result: " + JSON.stringify(result));
-              console.log("discarding unexpected result: " + JSON.stringify(result))
-              return;
+          var hasFirstResult = false;
+          db.executeSql("SELECT UPPER('first') as uppertext", [], function(rs) {
+            // EXPECTED ONCE:
+            expect(hasFirstResult).toBe(false);
+            hasFirstResult = true;
+            expect(rs).toBeDefined();
+            if (!!rs) {
+              expect(rs.rows).toBeDefined();
+              expect(rs.rows.length).toBe(1);
+              expect(rs.rows.item(0).uppertext).toBe('FIRST');
             }
+          }, function(error) {
+            // NOT EXPECTED:
+            expect(false).toBe(true);
+            expect(error.message).toBe('--');
+          });
 
-            ok(!!result, "valid result object");
-
-            // ignore cb (and do not count) if result is undefined:
-            if (!!result) {
-              console.log("result.rows.item(0).uppertext: " + result.rows.item(0).uppertext);
-              equal(result.rows.item(0).uppertext, expected[i], "Check result " + i);
-              i++;
-              start(1);
+          var hasSecondResult = false;
+          db.executeSql("SELECT UPPER('second') as uppertext", [], function(rs) {
+            // EXPECTED ONCE:
+            expect(hasFirstResult).toBe(true);
+            expect(hasSecondResult).toBe(false);
+            hasSecondResult = true;
+            expect(rs).toBeDefined();
+            if (!!rs) {
+              expect(rs.rows).toBeDefined();
+              expect(rs.rows.length).toBe(1);
+              expect(rs.rows.item(0).uppertext).toBe('SECOND');
             }
-          };
-
-          db.executeSql("select upper('first') as uppertext", [], okcb);
-          db.executeSql("select upper('second') as uppertext", [], okcb);
-        });
+          }, function(error) {
+            // NOT EXPECTED:
+            expect(false).toBe(true);
+            expect(error.message).toBe('--');
+          });
+        }, MYTIMEOUT);
 
       });
     }
@@ -133,13 +132,6 @@ var mytests = function() {
   });
 
   describe('Plugin: plugin-specific error test(s)', function() {
-
-    var pluginScenarioList = [
-      isAndroid ? 'Plugin-implementation-default' : 'Plugin',
-      'Plugin-implementation-2'
-    ];
-
-    var pluginScenarioCount = isAndroid ? 2 : 1;
 
     for (var i=0; i<pluginScenarioCount; ++i) {
 
@@ -150,10 +142,6 @@ var mytests = function() {
 
         // NOTE: MUST be defined in function scope, NOT outer scope:
         var openDatabase = function(first, second, third, fourth, fifth, sixth) {
-          //if (!isImpl2) {
-          //  return window.sqlitePlugin.openDatabase(first, second, third, fourth, fifth, sixth);
-          //}
-
           var dbname, okcb, errorcb;
 
           if (first.constructor === String ) {
@@ -180,7 +168,7 @@ var mytests = function() {
           return window.sqlitePlugin.openDatabase(dbopts, okcb, errorcb);
         }
 
-        it(suiteName + "Multiple db.executeSql error result test", function(done) {
+        it(suiteName + 'Multiple db.executeSql error result test', function(done) {
           // NOTE: this test checks that for db.executeSql(), the error result
           // callback is called exactly once, with the proper result:
           var db = openDatabase("Multiple-DB-sql-error-result-test.db", "1.0", "Demo", DEFAULT_SIZE);
@@ -191,9 +179,17 @@ var mytests = function() {
           db.executeSql("SELCT upper('first') AS uppertext", [], function() {
             // NOT EXPECTED:
             expect(false).toBe(true);
-          }, function(e) {
-            expect(e).toBeDefined();
-            // FUTURE TBD check error fields
+          }, function(error) {
+            // EXPECTED RESULT 1:
+            expect(error).toBeDefined();
+            expect(error.code).toBeDefined();
+            expect(error.message).toBeDefined();
+
+            // Check syntax error code/message:
+            if (!isWindows && !isWP8) // BROKEN (INCONSISTENT) on Windows:
+              expect(error.code).toBe(5);
+            if (!isWindows && !isWP8) // BROKEN (INCONSISTENT) on Windows:
+              expect(error.message).toMatch(/syntax error/);
 
             // CHECK that this was not called before
             expect(error_result_count).toBe(0);
@@ -204,9 +200,12 @@ var mytests = function() {
           db.executeSql("SELECT uper('second') as uppertext", [], function() {
             // NOT EXPECTED:
             expect(false).toBe(true);
-          }, function(e) {
-            expect(e).toBeDefined();
-            // FUTURE TBD check error fields
+          }, function(error) {
+            // EXPECTED RESULT 2:
+            expect(error).toBeDefined();
+            expect(error).toBeDefined();
+            expect(error.code).toBeDefined();
+            expect(error.message).toBeDefined();
 
             expect(error_result_count).toBe(1);
             ++error_result_count;
@@ -214,7 +213,7 @@ var mytests = function() {
             // and finish this test:
             done();
           });
-        });
+        }, MYTIMEOUT);
 
       });
     }
@@ -222,13 +221,6 @@ var mytests = function() {
   });
 
   describe('Plugin: more plugin-specific test(s)', function() {
-
-    var pluginScenarioList = [
-      isAndroid ? 'Plugin-implementation-default' : 'Plugin',
-      'Plugin-implementation-2'
-    ];
-
-    var pluginScenarioCount = isAndroid ? 2 : 1;
 
     for (var i=0; i<pluginScenarioCount; ++i) {
 
@@ -239,9 +231,6 @@ var mytests = function() {
 
         // NOTE: MUST be defined in function scope, NOT outer scope:
         var openDatabase = function(first, second, third, fourth, fifth, sixth) {
-          //if (!isImpl2) {
-          //  return window.sqlitePlugin.openDatabase(first, second, third, fourth, fifth, sixth);
-          //}
 
           var dbname, okcb, errorcb;
 
@@ -269,26 +258,30 @@ var mytests = function() {
           return window.sqlitePlugin.openDatabase(dbopts, okcb, errorcb);
         }
 
-        test_it(suiteName + "PRAGMAs & multiple database transactions mixed together", function() {
+        it(suiteName + "PRAGMAs & multiple database transactions mixed together", function(done) {
           var db = openDatabase("DB1", "1.0", "Demo", DEFAULT_SIZE);
 
           var db2 = openDatabase("DB2", "1.0", "Demo", DEFAULT_SIZE);
 
-          stop(2);
+          //stop(2);
 
+          // NOTE: This transaction will finish before the transaction started the db2 transaction:
           db.transaction(function(tx) {
             tx.executeSql('DROP TABLE IF EXISTS test_table');
             tx.executeSql('CREATE TABLE IF NOT EXISTS test_table (id integer primary key, data text, data_num integer)', [], function() {
               console.log("test_table created");
             });
 
-            stop();
+            //stop();
             db.executeSql("pragma table_info (test_table);", [], function(res) {
-              start();
+              //start();
               console.log("PRAGMA res: " + JSON.stringify(res));
-              equal(res.rows.item(2).name, "data_num", "DB1 table number field name");
+              // DB1 table number field name
+              expect(res.rows.item(2).name).toBe('data_num');
             });
           });
+
+          var waiting = 2;
 
           db2.transaction(function(tx) {
             tx.executeSql('DROP TABLE IF EXISTS tt2');
@@ -298,20 +291,28 @@ var mytests = function() {
 
             db.executeSql("pragma table_info (test_table);", [], function(res) {
               console.log("PRAGMA (db) res: " + JSON.stringify(res));
-              equal(res.rows.item(0).name, "id", "DB1 table key field name");
-              equal(res.rows.item(1).name, "data", "DB1 table text field name");
-              equal(res.rows.item(2).name, "data_num", "DB1 table number field name");
+              // DB1 table key field name
+              expect(res.rows.item(0).name).toBe('id');
+              // DB1 table text field name
+              expect(res.rows.item(1).name).toBe('data');
+              // DB1 table number field name
+              expect(res.rows.item(2).name).toBe('data_num');
 
-              start();
+              //start();
+              if (--waiting === 0) done();
             });
 
             db2.executeSql("pragma table_info (tt2);", [], function(res) {
               console.log("PRAGMA (tt2) res: " + JSON.stringify(res));
-              equal(res.rows.item(0).name, "id2", "DB2 table key field name");
-              equal(res.rows.item(1).name, "data2", "DB2 table text field name");
-              equal(res.rows.item(2).name, "data_num2", "DB2 table number field name");
+              // DB2 table key field name"
+              expect(res.rows.item(0).name).toBe('id2');
+              // DB2 table text field name
+              expect(res.rows.item(1).name).toBe('data2');
+              // DB2 table number field name
+              expect(res.rows.item(2).name).toBe('data_num2');
 
-              start();
+              //start();
+              if (--waiting === 0) done();
             });
           });
         });