+++ spec/www/spec/db-tx-error-mapping-test.js
@@ -4,44 +4,21 @@ var MYTIMEOUT = 12000;
 
 var DEFAULT_SIZE = 5000000; // max to avoid popup in safari/ios
 
-// FUTURE TODO replace in test(s):
-function ok(test, desc) { expect(test).toBe(true); }
-
-// XXX TODO REFACTOR OUT OF OLD TESTS:
-var wait = 0;
-var test_it_done = null;
-function xtest_it(desc, fun) { xit(desc, fun); }
-function test_it(desc, fun) {
-  wait = 0;
-  it(desc, function(done) {
-    test_it_done = done;
-    fun();
-  }, MYTIMEOUT);
-}
-function stop(n) {
-  if (!!n) wait += n
-  else ++wait;
-}
-function start(n) {
-  if (!!n) wait -= n;
-  else --wait;
-  if (wait == 0) test_it_done();
-}
-
 var isWP8 = /IEMobile/.test(navigator.userAgent); // Matches WP(7/8/8.1)
 var isWindows = /Windows /.test(navigator.userAgent); // Windows
 var isAndroid = !isWindows && /Android/.test(navigator.userAgent);
 
-// NOTE: In the common storage-master branch there is no difference between the
-// default implementation and implementation #2. But the test will also apply
-// the androidLockWorkaround: 1 option in the case of implementation #2.
+// NOTE: In certain versions such as Cordova-sqlcipher-adapter there is
+// no difference between the default implementation and implementation #2.
+// But the test will also specify the androidLockWorkaround: 1 option
+// in case of implementation #2 (also ignored by Cordova-sqlcipher-adapter).
 var scenarioList = [
   isAndroid ? 'Plugin-implementation-default' : 'Plugin',
   'HTML5',
   'Plugin-implementation-2'
 ];
 
-var scenarioCount = (!!window.hasWebKitBrowser) ? (isAndroid ? 3 : 2) : 1;
+var scenarioCount = (!!window.hasBrowserWithWebSQL) ? (isAndroid ? 3 : 2) : 1;
 
 var mytests = function() {
 
@@ -71,32 +48,36 @@ var mytests = function() {
         }
       }
 
+      describe(suiteName + 'SQLITE error code/message mapping test(s)', function() {
 
-        test_it(suiteName + "syntax error", function() {
+        it(suiteName + 'syntax error (incorrect error code & inconsistent message on Windows)', function(done) {
           var db = openDatabase("Syntax-error-test.db", "1.0", "Demo", DEFAULT_SIZE);
-          ok(!!db, "db object");
+          expect(db).toBeDefined();
 
-          stop(2);
+          var sqlerror = null; // VERIFY this was received
+
           db.transaction(function(tx) {
             tx.executeSql('DROP TABLE IF EXISTS test_table');
             tx.executeSql('CREATE TABLE IF NOT EXISTS test_table (data unique)');
 
-            // This insertion has a sql syntax error
+            // This insertion has a SQL syntax error
             tx.executeSql("insert into test_table (data) VALUES ", [123], function(tx) {
-              ok(false, "unexpected success");
-              start();
+              // NOT EXPECTED:
+              expect(false).toBe(true);
               throw new Error('abort tx');
+
             }, function(tx, error) {
+              sqlerror = error;
               expect(error).toBeDefined();
               expect(error.code).toBeDefined();
               expect(error.message).toBeDefined();
 
-              // BROKEN on Windows/WP8:
+              // Plugin BROKEN on Windows: INCORRECT error code (0: UNKNOWN_ERR)
               if (!isWindows && !isWP8)
                 expect(error.code).toBe(5);
 
-              // BROKEN (INCONSISTENT) on Windows (OK on wp8 platform):
-              if (!isWindows)
+              // Plugin BROKEN (INCONSISTENT) on Windows:
+              if (!isWindows && !isWP8)
                 expect(error.message).toMatch(/near .*\"*\"*:*syntax error/);
 
               // From built-in Android database exception message:
@@ -107,49 +88,57 @@ var mytests = function() {
               if (isWebSql)
                 expect(error.message).toMatch(/1 near .*\"*\"*:*syntax error/);
 
-              start();
-
-              // We want this error to fail the entire transaction
+              // FAIL transaction & check reported transaction error:
               return true;
             });
           }, function (error) {
-            ok(!!error, "valid error object");
-            ok(error.hasOwnProperty('message'), "error.message exists");
-            start();
+            expect(!!sqlerror).toBe(true); // VERIFY the SQL error callback was triggered
+
+            expect(error).toBeDefined();
+            expect(error.code).toBeDefined();
+            expect(error.message).toBeDefined();
+
+            isWebSql ? done() : db.close(done, done);
+          }, function() {
+            // NOT EXPECTED:
+            expect(false).toBe(true);
+            isWebSql ? done() : db.close(done, done);
           });
-        });
+        }, MYTIMEOUT);
 
-        test_it(suiteName + "constraint violation", function() {
+        it(suiteName + 'constraint violation (incorrect error code & inconsistent message on Windows)', function(done) {
           var db = openDatabase("Constraint-violation-test.db", "1.0", "Demo", DEFAULT_SIZE);
-          ok(!!db, "db object");
+          expect(db).toBeDefined();
 
-          stop(2);
+          var sqlerror = null; // VERIFY this was received
+
           db.transaction(function(tx) {
             tx.executeSql('DROP TABLE IF EXISTS test_table');
             tx.executeSql('CREATE TABLE IF NOT EXISTS test_table (data unique)');
 
-            tx.executeSql("insert into test_table (data) VALUES (?)", [123], null, function(tx, error) {
-              ok(false, error.message);
+            tx.executeSql('INSERT INTO test_table (data) VALUES (?)', [123], null, function(tx, error) {
+              expect(false).toBe(true);
+              expect(error.message).toBe('--');
             });
 
             // This insertion will violate the unique constraint
-            tx.executeSql("insert into test_table (data) VALUES (?)", [123], function(tx) {
-              ok(false, "unexpected success");
-              ok(!!res['rowsAffected'] || !(res.rowsAffected >= 1), "should not have positive rowsAffected");
-              start();
+            tx.executeSql('INSERT INTO test_table (data) VALUES (?)', [123], function(tx) {
+              // NOT EXPECTED:
+              expect(false).toBe(true);
               throw new Error('abort tx');
             }, function(tx, error) {
+              sqlerror = error;
               expect(error).toBeDefined();
               expect(error.code).toBeDefined();
               expect(error.message).toBeDefined();
 
-              // BROKEN on Windows/WP8:
-              if (!isWindows && !isWP8)
+              // Plugin BROKEN on Windows: INCORRECT error code (0: UNKNOWN_ERR)
+              if (!isWindows)
                 expect(error.code).toBe(6);
 
               if (isWebSql) // WebSQL may have a missing 'r' (iOS):
                 expect(error.message).toMatch(/constr?aint fail/);
-              else if (!isWindows && !isWP8) // BROKEN (INCONSISTENT) on Windows/WP8
+              else if (!isWindows && !isWP8) // Plugin BROKEN (INCONSISTENT) on Windows
                 expect(error.message).toMatch(/constraint fail/);
 
               // From built-in Android database exception message:
@@ -160,21 +149,29 @@ var mytests = function() {
               if (isWebSql && isAndroid)
                 expect(error.message).toMatch(/19 .*constraint fail/);
 
-              // SQLite error code part of Web SQL error.message (iOS):
+              // SQLite error code is apparently part of Web SQL error.message (iOS only):
               if (isWebSql && !isAndroid)
                 expect(error.message).toMatch(/19 .*not unique/);
 
-              start();
-
-              // We want this error to fail the entire transaction
+              // FAIL transaction & check reported transaction error:
               return true;
             });
-          }, function(error) {
-            ok(!!error, "valid error object");
-            ok(error.hasOwnProperty('message'), "error.message exists");
-            start();
+          }, function (error) {
+            expect(!!sqlerror).toBe(true); // VERIFY the SQL error callback was triggered
+
+            expect(error).toBeDefined();
+            expect(error.code).toBeDefined();
+            expect(error.message).toBeDefined();
+
+            isWebSql ? done() : db.close(done, done);
+          }, function() {
+            // NOT EXPECTED:
+            expect(false).toBe(true);
+            isWebSql ? done() : db.close(done, done);
           });
-        });
+        }, MYTIMEOUT);
+
+      });
 
     });
   }