+++ spec/www/spec/ext-tx-blob-test.js
@@ -7,26 +7,21 @@ var DEFAULT_SIZE = 5000000; // max to avoid popup in safari/ios
 // FUTURE TODO replace in test(s):
 function ok(test, desc) { expect(test).toBe(true); }
 
-var isAndroid = /Android/.test(navigator.userAgent);
 var isWP8 = /IEMobile/.test(navigator.userAgent); // Matches WP(7/8/8.1)
-//var isWindows = /Windows NT/.test(navigator.userAgent); // Windows [NT] (8.1)
-var isWindows = /Windows /.test(navigator.userAgent); // Windows (8.1)
-//var isWindowsPC = /Windows NT/.test(navigator.userAgent); // Windows [NT] (8.1)
-//var isWindowsPhone_8_1 = /Windows Phone 8.1/.test(navigator.userAgent); // Windows Phone 8.1
-//var isIE = isWindows || isWP8 || isWindowsPhone_8_1;
-var isIE = isWindows || isWP8;
-var isWebKit = !isIE; // TBD [Android or iOS]
+var isWindows = /Windows /.test(navigator.userAgent); // Windows 8.1/Windows Phone 8.1/Windows 10
+var isAndroid = !isWindows && /Android/.test(navigator.userAgent);
 
-// NOTE: In the core-master branch there is no difference between the default
-// implementation and implementation #2. But the test will also apply
-// the androidLockWorkaround: 1 option in the case of implementation #2.
+// NOTE: In certain versions such as Cordova-sqlcipher-adapter there is
+// no difference between the default implementation and implementation #2.
+// But the test will also specify the androidLockWorkaround: 1 option
+// in case of implementation #2 (also ignored by Cordova-sqlcipher-adapter).
 var scenarioList = [
   isAndroid ? 'Plugin-implementation-default' : 'Plugin',
   'HTML5',
   'Plugin-implementation-2'
 ];
 
-var scenarioCount = (!!window.hasWebKitBrowser) ? (isAndroid ? 3 : 2) : 1;
+var scenarioCount = (!!window.hasBrowserWithWebSQL) ? (isAndroid ? 3 : 2) : 1;
 
 var mytests = function() {
 
@@ -57,8 +52,8 @@ var mytests = function() {
 
       //describe(scenarioList[i] + ': tx blob test(s)', function() {
 
-        // XXX ENABLED for iOS ONLY (for now):
-        // This test shows that the plugin does not throw an error when trying to serialize
+        // ENABLED for iOS ONLY (for now):
+        // This test shows that the plugin does not throw an error when attempting to serialize
         // a non-standard parameter type. Blob becomes an empty dictionary on iOS, for example,
         // and so this verifies the type is converted to a string and continues. Web SQL does
         // the same but on the JavaScript side and converts to a string like `[object Blob]`.
@@ -72,7 +67,6 @@ var mytests = function() {
           // abort the test if ArrayBuffer is undefined
           // TODO: consider trying this for multiple non-standard parameter types instead
           if (typeof ArrayBuffer === "undefined") pending('ArrayBuffer type does not exist');
-
 
           var db = openDatabase("Blob-test.db", "1.0", "Demo", DEFAULT_SIZE);
           ok(!!db, "db object");
@@ -92,14 +86,19 @@ var mytests = function() {
             tx.executeSql('DROP TABLE IF EXISTS test_table');
             tx.executeSql('CREATE TABLE IF NOT EXISTS test_table (foo blob)');
             tx.executeSql('INSERT INTO test_table VALUES (?)', [blob], function(tx, res) {
-              ok(true, "INSERT blob OK");
+              // EXPECTED RESULT: INSERT Blob object OK
+              expect(true).toBe(true);
               done();
             }, function(tx, error) {
-              ok(false, "INSERT blob FAILED");
-              done();
+              // NOT EXPECTED: INSERT Blob object FAILED
+              expect(false).toBe(true);
+              expect(error.message).toBe('--');
+              return done() || false;
             });
-          }, function(err) { 
-            ok(false, "transaction failure with message: " + err.message);
+          }, function(error) {
+            // NOT EXPECTED: transaction failure
+            expect(false).toBe(true);
+            expect(error.message).toBe('--');
             done();
           });
         });