+++ spec/www/spec/regexp-test.js
@@ -10,11 +10,18 @@ var isWindows = /Windows /.test(navigator.userAgent); // Windows
 var isAndroidUA = /Android/.test(navigator.userAgent);
 var isAndroid = (isAndroidUA && !isWindows);
 
-var scenarioList = [ isAndroid ? 'Plugin-implementation-default' : 'Plugin', 'HTML5', 'Plugin-implementation-2' ];
+// NOTE: In certain versions such as Cordova-sqlcipher-adapter there is
+// no difference between the default implementation and implementation #2.
+// But the test will also specify the androidLockWorkaround: 1 option
+// in case of implementation #2 (also ignored by Cordova-sqlcipher-adapter).
+var scenarioList = [
+  isAndroid ? 'Plugin-implementation-default' : 'Plugin',
+  'HTML5',
+  'Plugin-implementation-2'
+];
 
-var scenarioCount = (!!window.hasWebKitBrowser) ? (isAndroid ? 3 : 2) : 1;
+var scenarioCount = (!!window.hasBrowserWithWebSQL) ? (isAndroid ? 3 : 2) : 1;
 
-// simple tests:
 var mytests = function() {
 
   for (var i=0; i<scenarioCount; ++i) {
@@ -43,7 +50,7 @@ var mytests = function() {
           if (isWP8) pending('NOT IMPLEMENTED for WP8');
           if (isWindows) pending('NOT IMPLEMENTED for Windows');
           //if (!isWebSql && isAndroid && isImpl2 && /Android [1-4]/.test(navigator.userAgent)) pending('BROKEN for android.database (version 1.x-4.x)');
-          if (!isWebSql && isAndroid) pending('SKIP for android.database'); // TBD (SKIP for Android plugin for now)
+          if (!isWebSql && isAndroid) pending('SKIP for android.database'); // TBD (SKIP for android.database [Android plugin] for now)
           if (!isAndroid && !isWindows && !isWP8) pending('SKIP for iOS');
 
           var db = openDatabase('simple-regexp-test.db', '1.0', 'test', DEFAULT_SIZE);